Kt 5 (#234)
* dodato oporezivanja, i jos neke izmene (#225)

* dodate neke izmene (#226)

* dodati soketi za myOption i geter (#227)

* odradjen email service (#228)

* odradjen email service

* zakomentarisan test u emailu

* Otc banka4 izmene konacne (#229)

* dodate rute za otc

* promenjen link za dev granu

* izmenjen test

* izmene za ogija (#230)

* Uradjeni svi testovi (#232)


* izmenjeni linkovi i actuary service (#233)

---------

+++ exchange-service/src/main/java/rs/edu/raf/exchangeservice/service/listingService/FutureService.java
@@ -51,9 +51,10 @@ public class FutureService {
                     future.setMaintenanceMargin(Integer.parseInt(parts[3]));
                     future.setType(parts[4]);
                     Random random = new Random();
-                    double min = 10000;
-                    double max = 50000;
-                    double price = random.nextDouble((max - min) + 1) + min;
+                    int min = 10000;
+                    int max = 50000;
+                    double price = random.nextInt((max - min) + 1) + min;
+
                     future.setPrice(price);
                     future.setCurrencyMark("RSD");
